Change every float with double

+++ problems/D6485-LD01_23-Arithmetic and Operation-JKT/problemAA.cpp
@@ -5,7 +5,7 @@ int main()
 	int assignment, mid, final;
 	scanf("%d %d %d", &assignment, &mid, &final);
 	
-	float finalScore = 20*assignment/100 + 30*mid/100 + 50*final/100;
+	double finalScore = 20*assignment/100 + 30*mid/100 + 50*final/100;
 	
-	printf("%.2f\n", finalScore);
+	printf("%.2lf\n", finalScore);
 }
